+++ src/client/rtp.rs
@@ -4,14 +4,46 @@
 //! RTP and RTCP handling; see [RFC 3550](https://datatracker.ietf.org/doc/html/rfc3550).
 
 use bytes::Bytes;
-use log::debug;
+use log::{debug, warn};
 
 use crate::client::PacketItem;
 use crate::rtcp::ReceivedCompoundPacket;
 use crate::rtp::{RawPacket, ReceivedPacket};
-use crate::{ConnectionContext, Error, ErrorInt, PacketContext, StreamContext, StreamContextInner};
+use crate::{
+    ConnectionContext, Error, ErrorInt, PacketContext, PacketContextInner, StreamContext,
+    StreamContextInner,
+};
 
-use super::{SessionOptions, Timeline};
+use super::{SessionOptions, Timeline, UnknownRtcpSsrcPolicy};
+
+/// Describes how Retina formed its initial expectation for the stream's `ssrc` or `seq`.
+#[derive(Copy, Clone, Debug)]
+enum InitialExpectation {
+    PlayResponseHeader,
+    RtpPacket,
+    RtcpPacket,
+}
+
+#[derive(Copy, Clone)]
+struct Ssrc {
+    init: InitialExpectation,
+    ssrc: u32,
+}
+
+impl std::fmt::Debug for Ssrc {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
+        f.debug_struct("Ssrc")
+            .field("init", &self.init)
+            .field("ssrc", &format_args!("0x{:08x}", self.ssrc))
+            .finish()
+    }
+}
+
+#[derive(Copy, Clone, Debug)]
+struct Seq {
+    init: InitialExpectation,
+    next: u16,
+}
 
 /// RTP/RTCP demarshaller which ensures packets have the correct SSRC and
 /// monotonically increasing SEQ. Unstable; exposed for benchmark.
@@ -34,23 +66,60 @@ use super::{SessionOptions, Timeline};
 #[doc(hidden)] // pub only for the benchmarks; not a stable API.
 #[derive(Debug)]
 pub struct InorderParser {
-    ssrc: Option<u32>,
-    next_seq: Option<u16>,
+    ssrc: Option<Ssrc>,
+    seq: Option<Seq>,
 
-    /// True iff `ssrc` was set from the beginning via a `RTP-Info` header.
-    initial_ssrc: bool,
+    /// Total RTP packets seen in this stream.
+    seen_rtp_packets: u64,
 
-    /// Total packets seen in this stream.
-    seen_packets: u64,
+    /// Total RTCP packets seen in this stream.
+    seen_rtcp_packets: u64,
+
+    unknown_rtcp_session: UnknownRtcpSsrcPolicy,
+    seen_unknown_rtcp_session: bool,
+}
+
+fn note_stale_live555_data_if_tcp(
+    tool: Option<&super::Tool>,
+    session_options: &SessionOptions,
+    conn_ctx: &crate::ConnectionContext,
+    stream_ctx: &StreamContext,
+    pkt_ctx: &PacketContext,
+) {
+    if let (
+        StreamContext(StreamContextInner::Tcp(stream_ctx)),
+        PacketContext(PacketContextInner::Tcp { msg_ctx }),
+    ) = (stream_ctx, pkt_ctx)
+    {
+        super::note_stale_live555_data(
+            tool,
+            session_options,
+            conn_ctx,
+            stream_ctx.rtp_channel_id,
+            msg_ctx,
+        );
+    }
 }
 
 impl InorderParser {
-    pub fn new(ssrc: Option<u32>, next_seq: Option<u16>) -> Self {
+    pub fn new(
+        ssrc: Option<u32>,
+        next_seq: Option<u16>,
+        unknown_rtcp_session: UnknownRtcpSsrcPolicy,
+    ) -> Self {
         Self {
-            ssrc,
-            next_seq,
-            initial_ssrc: ssrc.is_some(),
-            seen_packets: 0,
+            ssrc: ssrc.map(|ssrc| Ssrc {
+                init: InitialExpectation::PlayResponseHeader,
+                ssrc,
+            }),
+            seq: next_seq.map(|next| Seq {
+                init: InitialExpectation::PlayResponseHeader,
+                next,
+            }),
+            seen_rtp_packets: 0,
+            seen_rtcp_packets: 0,
+            unknown_rtcp_session,
+            seen_unknown_rtcp_session: false,
         }
     }
 
@@ -73,8 +142,8 @@ impl InorderParser {
                 pkt_ctx: *pkt_ctx,
                 stream_id,
                 description: format!(
-                    "corrupt RTP header while expecting seq={:04x?}: {:?}\n{:#?}",
-                    &self.next_seq,
+                    "corrupt RTP header while expecting seq={:?}: {:?}\n{:#?}",
+                    &self.seq,
                     e.reason,
                     crate::hex::LimitedHex::new(&e.data[..], 64),
                 ),
@@ -94,35 +163,27 @@ impl InorderParser {
 
         let sequence_number = raw.sequence_number();
         let ssrc = raw.ssrc();
-        let loss = sequence_number.wrapping_sub(self.next_seq.unwrap_or(sequence_number));
-        if matches!(self.ssrc, Some(s) if s != ssrc) {
-            if matches!(stream_ctx.0, StreamContextInner::Udp(_)) {
-                super::note_stale_live555_data(tool, session_options);
-            }
-            if self.seen_packets > 0 {
-                if false {
-                    bail!(ErrorInt::RtpPacketError {
-                        conn_ctx: *conn_ctx,
-                        pkt_ctx: *pkt_ctx,
-                        stream_ctx: stream_ctx.to_owned(),
-                        stream_id,
-                        ssrc,
-                        sequence_number,
-                        description: format!(
-                            "Wrong ssrc after {} packets; expecting ssrc={:08x?} seq={:04x?} \
-                             (initial ssrc: {:?})",
-                            self.seen_packets, self.ssrc, self.next_seq, self.initial_ssrc,
-                        ),
-                    });
-                } else {
-                    log::info!(
-                        "Skipping invalid rtp packet: Wrong ssrc after {} packets; expecting ssrc={:08x?} seq={:04x?} \
-                             (initial ssrc: {:?})",
-                            self.seen_packets, self.ssrc, self.next_seq, self.initial_ssrc,
-                    );
-                    return Ok(None);
-                }
-            }
+        let loss =
+            sequence_number.wrapping_sub(self.seq.map(|s| s.next).unwrap_or(sequence_number));
+        if matches!(self.ssrc, Some(s) if s.ssrc != ssrc) {
+            note_stale_live555_data_if_tcp(tool, session_options, conn_ctx, stream_ctx, pkt_ctx);
+            bail!(ErrorInt::RtpPacketError {
+                conn_ctx: *conn_ctx,
+                pkt_ctx: *pkt_ctx,
+                stream_ctx: stream_ctx.to_owned(),
+                stream_id,
+                ssrc,
+                sequence_number,
+                description: format!(
+                    "wrong ssrc after {} RTP pkts + {} RTCP pkts; expecting ssrc={:?} seq={:?}",
+                    self.seen_rtp_packets, self.seen_rtcp_packets, self.ssrc, self.seq,
+                ),
+            });
+        } else if self.ssrc.is_none() {
+            self.ssrc = Some(Ssrc {
+                init: InitialExpectation::RtpPacket,
+                ssrc,
+            });
         }
         if loss > 0x80_00 {
             if matches!(stream_ctx.0, StreamContextInner::Tcp { .. }) {
@@ -134,23 +195,23 @@ impl InorderParser {
                     ssrc,
                     sequence_number,
                     description: format!(
-                        "Out-of-order packet or large loss; expecting ssrc={:08x?} seq={:04x?}",
-                        self.ssrc, self.next_seq
+                        "Out-of-order packet or large loss; expecting ssrc={:08x?} seq={:?}",
+                        self.ssrc, self.seq
                     ),
                 });
             } else {
                 log::info!(
-                    "Skipping out-of-order seq={:04x} when expecting ssrc={:08x?} seq={:04x?}",
+                    "Skipping out-of-order seq={} when expecting ssrc={:08x?} seq={:?}",
                     sequence_number,
                     self.ssrc,
-                    self.next_seq,
+                    self.seq,
                 );
                 return Ok(None);
             }
         }
         let timestamp = match timeline.advance_to(raw.timestamp()) {
             Ok(ts) => ts,
-            Err(description) if false => bail!(ErrorInt::RtpPacketError {
+            Err(description) => bail!(ErrorInt::RtpPacketError {
                 conn_ctx: *conn_ctx,
                 pkt_ctx: *pkt_ctx,
                 stream_ctx: stream_ctx.to_owned(),
@@ -159,17 +220,15 @@ impl InorderParser {
                 sequence_number,
                 description,
             }),
-            Err(description) if true => {
-                log::info!(
-                    "Failed to advance timestamp: {}", description,
-                );
-                return Ok(None);
-            },
-            _ => unreachable!()
         };
-        self.ssrc = Some(ssrc);
-        self.next_seq = Some(sequence_number.wrapping_add(1));
-        self.seen_packets += 1;
+        self.seq = Some(Seq {
+            init: self
+                .seq
+                .map(|s| s.init)
+                .unwrap_or(InitialExpectation::RtpPacket),
+            next: sequence_number.wrapping_add(1),
+        });
+        self.seen_rtp_packets += 1;
         Ok(Some(PacketItem::Rtp(ReceivedPacket {
             ctx: *pkt_ctx,
             stream_id,
@@ -186,6 +245,7 @@ impl InorderParser {
         session_options: &SessionOptions,
         stream_ctx: &StreamContext,
         tool: Option<&super::Tool>,
+        conn_ctx: &ConnectionContext,
         pkt_ctx: &PacketContext,
         timeline: &mut Timeline,
         stream_id: usize,
@@ -202,17 +262,46 @@ impl InorderParser {
             )?);
 
             let ssrc = sr.ssrc();
-            if matches!(self.ssrc, Some(s) if s != ssrc) {
-                if matches!(stream_ctx.0, StreamContextInner::Tcp { .. }) {
-                    super::note_stale_live555_data(tool, session_options);
+            if matches!(self.ssrc, Some(s) if s.ssrc != ssrc) {
+                match self.unknown_rtcp_session {
+                    UnknownRtcpSsrcPolicy::AbortSession => {
+                        note_stale_live555_data_if_tcp(
+                            tool,
+                            session_options,
+                            conn_ctx,
+                            stream_ctx,
+                            pkt_ctx,
+                        );
+                        return Err(format!(
+                            "Expected ssrc={:08x?}, got RTCP SR ssrc={:08x}",
+                            self.ssrc, ssrc
+                        ));
+                    }
+                    UnknownRtcpSsrcPolicy::Default | UnknownRtcpSsrcPolicy::DropPackets => {
+                        if !self.seen_unknown_rtcp_session {
+                            warn!(
+                                "saw unknown rtcp ssrc {ssrc}; rtp session has ssrc {s:?}",
+                                s = self.ssrc
+                            );
+                            self.seen_unknown_rtcp_session = true;
+                        }
+                        return Ok(None);
+                    }
+                    UnknownRtcpSsrcPolicy::ProcessPackets => {}
                 }
-                return Err(format!(
-                    "Expected ssrc={:08x?}, got RTCP SR ssrc={:08x}",
-                    self.ssrc, ssrc
-                ));
+            } else if self.ssrc.is_none()
+                && !matches!(
+                    self.unknown_rtcp_session,
+                    UnknownRtcpSsrcPolicy::ProcessPackets
+                )
+            {
+                self.ssrc = Some(Ssrc {
+                    init: InitialExpectation::RtcpPacket,
+                    ssrc,
+                });
             }
-            self.ssrc = Some(ssrc);
         }
+        self.seen_rtcp_packets += 1;
         Ok(Some(PacketItem::Rtcp(ReceivedCompoundPacket {
             ctx: *pkt_ctx,
             stream_id,
@@ -237,7 +326,7 @@ mod tests {
     #[test]
     fn geovision_pt50_packet() {
         let mut timeline = Timeline::new(None, 90_000, None).unwrap();
-        let mut parser = InorderParser::new(Some(0xd25614e), None);
+        let mut parser = InorderParser::new(Some(0xd25614e), None, UnknownRtcpSsrcPolicy::Default);
         let stream_ctx = StreamContext::dummy();
 
         // Normal packet.
@@ -292,7 +381,7 @@ mod tests {
     #[test]
     fn out_of_order() {
         let mut timeline = Timeline::new(None, 90_000, None).unwrap();
-        let mut parser = InorderParser::new(Some(0xd25614e), None);
+        let mut parser = InorderParser::new(Some(0xd25614e), None, UnknownRtcpSsrcPolicy::Default);
         let stream_ctx = StreamContext(StreamContextInner::Udp(UdpStreamContext {
             local_ip: IpAddr::V4(Ipv4Addr::UNSPECIFIED),
             peer_ip: IpAddr::V4(Ipv4Addr::UNSPECIFIED),